Replaced uint32_t with size_t.

+++ include/PolyVox/Mesh.h
@@ -55,7 +55,7 @@ namespace PolyVox
 		const VertexType* getRawVertexData(void) const;
 		POLYVOX_DEPRECATED const std::vector<VertexType>& getVertices(void) const;
 
-		uint32_t getNoOfIndices(void) const;
+		size_t getNoOfIndices(void) const;
 		IndexType getIndex(uint32_t index) const;
 		const IndexType* getRawIndexData(void) const;
 		POLYVOX_DEPRECATED const std::vector<IndexType>& getIndices(void) const;
